requirements.txt
argcomplete
beautifulsoup4
chardet
docx2txt
EbookLib
et-xmlfile
extract-msg
IMAPClient
jdcal
lxml
numpy
olefile
openpyxl
pandas
pdfminer.six
Pillow
pycryptodome
PyPDF2
python-dateutil
python-docx
python-pptx
pytz
six
sortedcontainers
soupsieve
SpeechRecognition
tesseract
textract
tzlocal
xlrd
XlsxWriter
yara-python
